Move new BST functions to unstable
