+++ data_structures/_binary_search_tree_internals.ts
@@ -24,8 +24,6 @@ export const internals: {
   findNode<T>(
     tree: BinarySearchTree<T>,
     value: T,
-    select?: "higher" | "lower",
-    returnIfFound?: boolean,
   ): BinarySearchNode<T> | null;
   rotateNode<T>(
     tree: BinarySearchTree<T>,
