+++ data_structures/deno.json
@@ -6,6 +6,7 @@
     "./unstable-bidirectional-map": "./unstable_bidirectional_map.ts",
     "./binary-heap": "./binary_heap.ts",
     "./binary-search-tree": "./binary_search_tree.ts",
+    "./unstable-binary-search-tree": "./unstable_binary_search_tree.ts",
     "./comparators": "./comparators.ts",
     "./red-black-tree": "./red_black_tree.ts"
   }
